Condition for disabling preview activates when mentioning users/etc
The condition to disable the bot previews checks if angular brackets meet _anywhere_ in the message, which ends up capturing mentions towards other users, roles, channels, or if custom emojis/animated nitro emojis are present anywhere in the message, as they respectively appear as `<@user_id>`, `<@&role_id>`, `<#channel_id>`, `<:emoji_id:>` and `<a:animated_emoji_id:>` in the `message.content`, all with meeting angular brackets.

I've changed `disable_command_pattern` to prevent returning a match if the character after the first bracket match any of the scenarios presented above. I don't have a running instance of the bot though so I haven't tested if it works but it should.